@@ -1,12 +1,11 @@
 ﻿using Unite.Data.Services.Configuration.Options;
 
-namespace Unite.Data.Migrations.Configuration.Options
+namespace Unite.Data.Migrations.Configuration.Options;
+
+public class SqlOptions : ISqlOptions
 {
-    public class SqlOptions : ISqlOptions
-    {
-        public string Host => EnvironmentConfig.SqlHost;
-        public string Port => EnvironmentConfig.SqlPort;
-        public string User => EnvironmentConfig.SqlUser;
-        public string Password => EnvironmentConfig.SqlPassword;
-    }
+    public string Host => EnvironmentConfig.SqlHost;
+    public string Port => EnvironmentConfig.SqlPort;
+    public string User => EnvironmentConfig.SqlUser;
+    public string Password => EnvironmentConfig.SqlPassword;
 }
