@@ -2,16 +2,15 @@
 using Unite.Data.Migrations.Configuration.Options;
 using Unite.Data.Services;
 
-namespace Unite.Data.Migrations
-{
-    public class DomainDbContextFactory : IDesignTimeDbContextFactory<DomainDbContext>
-    {
-        public DomainDbContext CreateDbContext(string[] args)
-        {
-            var options = new SqlOptions();
-            var dbContext = new DomainDbContext(options);
+namespace Unite.Data.Migrations;
 
-            return dbContext;
-        }
+public class DomainDbContextFactory : IDesignTimeDbContextFactory<DomainDbContext>
+{
+    public DomainDbContext CreateDbContext(string[] args)
+    {
+        var options = new SqlOptions();
+        var dbContext = new DomainDbContext(options);
+
+        return dbContext;
     }
 }
